requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Cors==3.0.10
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector==2.2.9
six==1.16.0
typing_extensions==4.5.0
Werkzeug==2.2.3
